refactor CNN utilities
