+++ hls4ml/converters/keras/convolution.py
@@ -1,10 +1,7 @@
 import math
 from hls4ml.converters.keras_to_hls import parse_default_keras_layer
 from hls4ml.converters.keras_to_hls import keras_handler
-from hls4ml.converters.keras_to_hls import parse_data_format
-from hls4ml.converters.keras_to_hls import compute_padding_1d
-from hls4ml.converters.keras_to_hls import compute_padding_2d
-
+from hls4ml.converters.utils import *
 
 @keras_handler('Conv1D', 'SeparableConv1D')
 def parse_conv1d_layer(keras_layer, input_names, input_shapes, data_reader, config):
