+++ hls4ml/converters/keras/reshape.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from hls4ml.converters.keras_to_hls import parse_default_keras_layer
 from hls4ml.converters.keras_to_hls import keras_handler
-from hls4ml.converters.keras_to_hls import parse_data_format
+from hls4ml.converters.utils import parse_data_format
 
 @keras_handler('Reshape')
 def parse_reshape_layer(keras_layer, input_names, input_shapes, data_reader, config):
